+++ IsometricEngine/item.h
@@ -22,11 +22,11 @@ public:
 	actor* owner;
 	glm::vec2 collisionBoundXY = glm::vec2(.1f), collisionBoundZ =glm::vec2(7.5f, 32.5f);
 	Texture2D UI_img;
-	float value = 100;
+	double value = 100;
 	int amount = 1;
 	bool autoPickup = true;
-	float pickupRange = .8;
+	double pickupRange = .8;
 	int item_ID;
-	float _internTimer = 0;
+	double _internTimer = 0;
 	
 };
