+++ IsometricEngine/textrenderer.h
@@ -43,7 +43,7 @@ public:
 	// pre-compiles a list of characters from the given font
 	void Load(std::string font, unsigned int fontSize);
 	// renders a string of text using the precompiled list of characters
-	void RenderText(std::string text, float x, float y, float scale, glm::vec3 color = glm::vec3(1.0f));
+	void RenderText(std::string text, double x, double y, double scale, glm::vec3 color = glm::vec3(1.0f));
 private:
 	// render state
 	unsigned int VAO, VBO;
